+++ db/migrate/20151130201513_add_community_manager_to_users.rb
@@ -1,0 +1,5 @@
+class AddCommunityManagerToUsers < ActiveRecord::Migration
+  def change
+    add_column :users, :community_manager, :boolean, null: false, default: false
+  end
+end