+++ test/controllers/blog_controller_test.rb
@@ -1,0 +1,18 @@
+require 'test_helper'
+
+class BlogControllerTest < ActionController::TestCase
+  test 'should get blog' do
+    get :blog
+    assert_response :success
+  end
+
+  test 'should show published blog' do
+    get :show, id: broadcasts(:published)
+    assert_response :success
+  end
+
+  test 'should not show draft blog' do
+    get :show, id: broadcasts(:draft)
+    assert_redirected_to blog_path
+  end
+end